style: change WhoWeAreContentLocalizationHandlerHelper location and move it to Helpers folder
